adk_testbed/requirements.txt
google-adk[eval]
pytest
pytest-asyncio
rouge-score
